Update a slot of your jenga stack

Update a slot of your jenga stack

update_stack(stack, objects, slots, save = TRUE)

Arguments

stack
A jenga stack as created by create.slot

objects
a list of items to be inserted into the jenga stack

slots
a vector of character inputs, of the slot/s to be updated. Must match slots already in the stack. Order of slots corresponds to order of objects

SAVE
if SAVE=TRUE, then the updated slot will be saved to disk in an .RData file of the name of the stack.

Value

an updated jenja stack